chore(example): dedup iOS/macOS test-only ExifKeyReader into one shared Swift file
The `readExifKeys` test-only method channel (added in #391/#392) was
implemented twice — once in `example/ios/Runner/AppDelegate.swift`
and once in `example/macos/Runner/MainFlutterWindow.swift`. Both used
the same `CGImageSourceCopyPropertiesAtIndex` walk with identical
prefix logic. iOS and macOS share ImageIO, so the code is byte-for-
byte identical modulo the Flutter/FlutterMacOS import.

Extract the implementation to `example/shared/ExifTestHelper.swift`
and reference the same file from both Runner Xcode projects via a
relative-path Xcode file reference (`sourceTree = SOURCE_ROOT`,
`path = ../shared/ExifTestHelper.swift`). The Runner's own Swift now
does one call: `ExifTestHelper.register(with: registrar.messenger())`.

Net: the reader logic exists in exactly one place. Adding a new
metadata dict to inspect requires editing one file, not two.
`#if canImport(Flutter) / FlutterMacOS` inside the shared file
handles the framework-import split without duplication.

Verified locally: iOS 26.5 simulator 18/18 pass, macOS host 13 pass
+ 5 iOS-only skipped — same as before this refactor.
